consolidate ngApp.js into index.js
Since these have both been simplified we can merge them together
